+++ src/interfaces/triggers/values/plugins/templates/ITemplateContext.php
@@ -1,5 +1,5 @@
 <?php
-namespace deflou\interfaces\triggers\operations\plugins\templates;
+namespace deflou\interfaces\triggers\values\plugins\templates;
 
 use extas\interfaces\IHasName;
 use extas\interfaces\IItem;
@@ -7,5 +7,5 @@ use extas\interfaces\parameters\IHaveParams;
 
 interface ITemplateContext extends IItem, IHasName, IHaveParams
 {
-    public const SUBJECT = 'deflou.trigger.operation.plugin.template';
+    public const SUBJECT = 'deflou.trigger.value.plugin.template';
 }